requirements.txt
flask
python-dotenv
anthropic
openai
pyairtable
gunicorn